+++ relay-server/src/statsd.rs
@@ -178,6 +178,10 @@ pub enum RelayHistograms {
     /// The distribution of buckets should be even.
     /// If it is not, this metric should expose it.
     PartitionKeys,
+
+    /// Measures how many transactions were created from segment spans in a single envelope.
+    #[cfg(feature = "processing")]
+    TransactionsFromSpansPerEnvelope,
 }
 
 impl HistogramMetric for RelayHistograms {
@@ -211,6 +215,10 @@ impl HistogramMetric for RelayHistograms {
             RelayHistograms::UpstreamEnvelopeBodySize => "upstream.envelope.body_size",
             RelayHistograms::UpstreamMetricsBodySize => "upstream.metrics.body_size",
             RelayHistograms::PartitionKeys => "metrics.buckets.partition_keys",
+            #[cfg(feature = "processing")]
+            RelayHistograms::TransactionsFromSpansPerEnvelope => {
+                "transactions_from_spans_per_envelope"
+            }
         }
     }
 }
@@ -642,6 +650,9 @@ pub enum RelayCounters {
     /// This metric is tagged with:
     /// - `decision`: "drop" if dynamic sampling drops the envelope, else "keep".
     DynamicSamplingDecision,
+    /// Counts how many transactions were created from segment spans.
+    #[cfg(feature = "processing")]
+    TransactionsFromSpans,
 }
 
 impl CounterMetric for RelayCounters {
@@ -680,6 +691,8 @@ impl CounterMetric for RelayCounters {
             RelayCounters::OpenTelemetryEvent => "event.opentelemetry",
             RelayCounters::GlobalConfigFetched => "global_config.fetch",
             RelayCounters::DynamicSamplingDecision => "dynamic_sampling_decision",
+            #[cfg(feature = "processing")]
+            RelayCounters::TransactionsFromSpans => "transactions_from_spans",
         }
     }
 }